Added basic page to display incident and its reports
Reports are currently grouped under a single Incident
(currently called 'Event' in the API).

This page shows the information about the Incident
and lists any nested Reports with their basic information.

Additional information will be displayed in a separate
Report page.
This is obviously subject to change but for the time being
I'm assuming we don't want to display everything everywhere
all at once.