Renamed incidents' pages URLs (and nunjuck template)
The names of these "things" are still in flux.
The API currently call these "things" events but this
is possibly subject to change and not necessarily the
right word for what they are.

At least for now and in the user-facing side of things
we'll stick with "incidents" on the UI side of things.

I'm keeping stuff coming from the API and adjacent
methods/variable as "events" (e.g. `getEvents()`
and `events` variables) as the endpoint and payload
still refers to "events".