IR-375: Placeholder pages to display an incident/a report (#131)
* Added basic page to display incident and its reports

Reports are currently grouped under a single Incident
(currently called 'Event' in the API).

This page shows the information about the Incident
and lists any nested Reports with their basic information.

Additional information will be displayed in a separate
Report page.
This is obviously subject to change but for the time being
I'm assuming we don't want to display everything everywhere
all at once.

* Added page to see single report

...linked from Incident page.

TODO: Show more details in this page (e.g. questions)

* When showing a report, display its details

including staff/prisoner involvement, questions/responses
and correction requests.

* Renamed incidents' pages URLs (and nunjuck template)

The names of these "things" are still in flux.
The API currently call these "things" events but this
is possibly subject to change and not necessarily the
right word for what they are.

At least for now and in the user-facing side of things
we'll stick with "incidents" on the UI side of things.

I'm keeping stuff coming from the API and adjacent
methods/variable as "events" (e.g. `getEvents()`
and `events` variables) as the endpoint and payload
still refers to "events".

* Fix indentation of incident/report templates

* Removed non-existent CSS class

* Show prisoner involvement outcome

as per PR suggestion (however there are probably
more things we could show or possibly hide)