IR-505: Skeleton of config validation
When updating the existing Incident Report types the
generated config is checked for validity.

If any errors for a given type config are detected
these are shown with a red "WARNING" message to highlight
the problem.

NOTE: The config file is written regardless of its validity.
This is by design as it makes it easier to inspect and git diff
against previous version.

Obviously if the update highlights some problems the updated
config should not be checked into Git unless you really know
what you're doing.