IR-505: Present inactive questions/answers differently
Inactive questions/answers are greyed out. It's now possible to see
inactive paths at a glance.

This is very useful when checking questionnaires for possible errors.

For example (see `RELEASED_IN_ERROR`) in a report type config some answers may lead to an inactive
question and peraphs there is another active question with the same label.
Was the answer supposed to lead to the equivalent active question instead?