IR-505: Validation honours active flags
Questions/answers which are not active are not added to the graph.

NOTE: Because of this change, if an answer leads to an inactive question
this will be detected as an invalid node (as it's not added to the graph).
For this reason the wording of the error has been tweaked to point out
the question may be unknown or it may be inactive.