IR-505: Highlight config is active when it has errors
When updating the Incident types configurations, if a config has problems
and it is active show a message saying it is active.

Example:

```
WARNING: Config for XYZ has the following errors:
 - config is active
 - some answers lead to these unknown or inactive questions: 61304
 - [...]
```

This is useful because some inactive config are currently invalid but
that's acceptable as they're not active.

Hovever if a config is currently active and has validation issues this
is more concerning.

With this change it is easier to spot which validation errors could be
ignored and which ones should be looked more carefully.