IR-506: Extended report type config to have allowed prisoner roles
Similarly to NOMIS config, new DPS config will have a list of allowed
prisoner involvement roles that could be used for this specific report
type.

This config is fairly straightforward, perhaps the most insteresting
feature is the `onlyOneAllowed` boolean field which can be used to prevent
having multiple prisoner involvements with the same role.
For example, maybe a given type can only allow only a single prisoner to be marked
as the `PERPETRATOR`.

NOTE: Generated report type configuration are now invalid as the conversion
doesn't populate this field just yet