IR-505: Config validation and visualisation (#184)
* IR-505: Skeleton of config validation

When updating the existing Incident Report types the
generated config is checked for validity.

If any errors for a given type config are detected
these are shown with a red "WARNING" message to highlight
the problem.

NOTE: The config file is written regardless of its validity.
This is by design as it makes it easier to inspect and git diff
against previous version.

Obviously if the update highlights some problems the updated
config should not be checked into Git unless you really know
what you're doing.

* IR-505: Added tests for validations

- starting question presence
- unreachable questions
- cycles (both positive and false positive/optional questions)

* IR-505: Check if questionnaire contains unknown questions

* IR-505: Checks for cycles and unreachable questions

* IR-505: Added validation test against complex config

Some WIP code/logic seems to have trouble with the `ESCAPE_FROM_CUSTODY`
questionnaire. Possibly because of its size or because it may contain
cycles.

Not clear yet but it's worth having a test to automatically validate
this report.

* Reworked DFS logic moved in its own generic module

NOTE: This is still subject to change/refactoring

* IR-505: Refactored Graph logic into a class

and moved more logic into this class

* IR-505: Document Graph methods

and tweaked cycle values returned to include node
that triggered the loop.

* IR-505: Generate Graphviz for a questionnaire config

* IR-505: Update report types saves graphviz files

Both in the DOT format and SVG format.

NOTE: These questionnaire are massive and viewing
them in the browser will probably not cut it.
I suggest using Inkskape (or similar vector graphics
program).

* IR-505: Refactoring: Moved graphviz conversion

Move the conversion logic (to build the DOT format) into the
'conversion' module where it belongs.

* IR-505: Refactored TS generation to split conversion/persistance

Separated logic to generate TS data for the questionnaire config from
the logic to save/format into a file.

* IR-505: Graphviz DOT files for Report types

Generated using `updateNomisIncidentTypeConfigurations` script

* IR-505: SVG files for Report types

Generated using `updateNomisIncidentTypeConfigurations` script.
These SVG files are converted from the DOT files.

* IR-505: Added active flags to questions/answers

The NOMIS questionnaires config have this information.

The alternative is to just lose this information which is not great as
NOMIS probably uses these flags to show/hide questions/answers.

NOTE: These are currently not taken in consideration by the validation
nor by the visualisation. They probably should.
I think this may account for some of the validation errors we're getting
with some of the report types.
I've seen some active questions pointing to inactive questions
(e.g. `RELEASED_IN_ERROR`) or even an inactive question being the
starting question (e.g. `OLD_FINDS4`) so there is definitely something
odd going on.

TODO: We should check how NOMIS currently works.

* IR-505: Present inactive questions/answers differently

Inactive questions/answers are greyed out. It's now possible to see
inactive paths at a glance.

This is very useful when checking questionnaires for possible errors.

For example (see `RELEASED_IN_ERROR`) in a report type config some answers may lead to an inactive
question and peraphs there is another active question with the same label.
Was the answer supposed to lead to the equivalent active question instead?

* IR-505: Validation honours active flags

Questions/answers which are not active are not added to the graph.

NOTE: Because of this change, if an answer leads to an inactive question
this will be detected as an invalid node (as it's not added to the graph).
For this reason the wording of the error has been tweaked to point out
the question may be unknown or it may be inactive.

* IR-505: display NOMIS code when updating report types

This is useful when investigating problems as it saves the trip to the
file with the mapping between old and new codes.

* IR-505: Check starting question is active

* IR-505: Highlight config is active when it has errors

When updating the Incident types configurations, if a config has problems
and it is active show a message saying it is active.

Example:

```
WARNING: Config for XYZ has the following errors:
 - config is active
 - some answers lead to these unknown or inactive questions: 61304
 - [...]
```

This is useful because some inactive config are currently invalid but
that's acceptable as they're not active.

Hovever if a config is currently active and has validation issues this
is more concerning.

With this change it is easier to spot which validation errors could be
ignored and which ones should be looked more carefully.

* IR-505: Concistency across error messages

Start all error messages with a lower case letter.

* IR-505: Check starting question is known

The starting question should be part of the questions set (`config.questions`)

* IR-505: Added note on where to download Graphviz cli

Graphviz's `dot` command is used by the `updateNomisIncidentTypeConfigurations`
script to generate SVG files for the questionnaires flows.

---------

Co-authored-by: Michael Willis <michael.willis@digital.justice.gov.uk>