IR-519: Starting to segregate WIP committed
We currently have a combination of more stable
code and code which is more "experimental"/exploratory.

Possibly most of the spikes code could be thrown away.

The idea is to have keep the stuff that we are currently
less sure about under `wip/` directory.
Over time we want to slowly move bits out of these
directories.