Tweaks to steps/fields generation
Backported some of the bits stuff currently in Josh's
implementation. Aiming to get a single version of these
functions that generate steps/fields (the one in
`server/data/incidentTypeConfiguration/formWizard.ts`).

Also setting the `controller` field as this is needed.

~I'm currently using the WIP `GenFieldsController` which
is also used in other places...~
I'm currently using the WIP `QuestionsController` which
is also used in other places...

but I haven't looked into
how the controllers works in much depth just yet.
~At the very least I'd like to rename this controller.~

NOTE: Some of the functionality, like showing comments/date
when `commentRequired`/`dateRequired` is true is currently
not here.