Generate steps (#209)
* Created function to generate all fields for a given incident type. Added a page using the form wizard to display generated fields

* moved page, controller and router into wip folders

* added function to generate steps from incident config and added to existing assaults example route

* generateFields() filter out inactive answers

* Renamed `GenFieldsController` to `QuestionsController`

Also, no need to have it's own directory AFAICT.

(I haven't touched the other controllers just yet
but I think they'd need some moving too...some of them
as obviously WIP/result of spike/experimentation
and should be moved to `/wip` as well)

* Tweaks to steps/fields generation

Backported some of the bits stuff currently in Josh's
implementation. Aiming to get a single version of these
functions that generate steps/fields (the one in
`server/data/incidentTypeConfiguration/formWizard.ts`).

Also setting the `controller` field as this is needed.

~I'm currently using the WIP `GenFieldsController` which
is also used in other places...~
I'm currently using the WIP `QuestionsController` which
is also used in other places...

but I haven't looked into
how the controllers works in much depth just yet.
~At the very least I'd like to rename this controller.~

NOTE: Some of the functionality, like showing comments/date
when `commentRequired`/`dateRequired` is true is currently
not here.

* Fixes to `/:reportType/questions` endpoint

- renamed/tweaked template to play well with how
  steps are generated. For example it's expecting
  the template to be named `questionPage` (and being
  under whatever directory was specified in the
  `templatePath` option)
- document "why" we currently have both `checkSession`
  and `csrf` set to `false`.
  - still not quite sure how the controller part of this
    work but at least we're highlighting the problem
    for future us

* Filter out inactive questions in steps generation

* Tweaks to `generateSteps()`

Hopefully this is a bit more readable. I'm keeping the
`next` array generation separate as it may be more
readable this way but more importantly because we
may add some logic to have less items when multiple
answers lead to the same question.

* Tidied up names/paths of `generate_(fields|steps)`

- moved these functions under `server/data/incidentTypeConfiguration/wip/`
closer to the other `generateSteps()`/`generateFields()`
- renamed to use `camelCase` as per JavaScript convetion
(me loves `snake_case`` tho!)
- renamed to `generateStepsWithConditionals()`/`generateFieldsWithConditionals()`
to highlight the main difference with the other functions
- adjusted import paths

---------

Co-authored-by: Aldo "xoen" Giambelluca <aldo.giambelluca@digital.justice.gov.uk>