Proof of concept macro to render field/conditionals
A lot of rough edges at the moment but hopefully the
general idea is sound.

Main concern is how we discriminate these conditional
fields, expecially when multiple answers can be selected,
for example:

```
Q: Where did it hit?
  [X] A: Neck
    Date [ DD/MM/YYYY]
    Comment [       ]
  [X] A: Torso
    Date [ DD/MM/YYYY]
    Comment [       ]
  [ ] A: Leg
  [ ] ...
```

How do we distinguish between neck's comment and
torso's comment?

At the moment I'm just generating an "id" like
"61293-Neck-comment" but this is clearly not ideal
and bound to cause all sort of problems (with spaces
and "odd" characters in the answers).

This is something that we'll have to handle somehow
but possibly we'll be determine as part of the controller
work.

NOTE: This macro is heavily WIP, only add conditional
to radio button (not checkboxes) and it could be tidied
(e.g. extract conditional field generation out)
and moved in a place where it can be shared and used
by other templates. But it's a start.