Moved field rendering logic into a partial (#223)
* Added plain vanilla `BaseController`

This is at the moment just a `FormWizard.Controller` with
no customisation at all. Just the upstream behaviour.

* Default `QuestionsController` uses vanilla `BaseController`

Interestingly things like (basic) validation and going
through pages still work.

* WIP: Explicitly render fields

instead of using `callAsMacro()`

* Added `dateRequired`/`commentRequired` in fields

This information can be used by the templates to
render the conditionals section of fields.

* Proof of concept macro to render field/conditionals

A lot of rough edges at the moment but hopefully the
general idea is sound.

Main concern is how we discriminate these conditional
fields, expecially when multiple answers can be selected,
for example:

```
Q: Where did it hit?
  [X] A: Neck
    Date [ DD/MM/YYYY]
    Comment [       ]
  [X] A: Torso
    Date [ DD/MM/YYYY]
    Comment [       ]
  [ ] A: Leg
  [ ] ...
```

How do we distinguish between neck's comment and
torso's comment?

At the moment I'm just generating an "id" like
"61293-Neck-comment" but this is clearly not ideal
and bound to cause all sort of problems (with spaces
and "odd" characters in the answers).

This is something that we'll have to handle somehow
but possibly we'll be determine as part of the controller
work.

NOTE: This macro is heavily WIP, only add conditional
to radio button (not checkboxes) and it could be tidied
(e.g. extract conditional field generation out)
and moved in a place where it can be shared and used
by other templates. But it's a start.

* Extracted `renderField` macro into its own file

Also removed unused import in `questionPage.njk`

* `renderField()` renders conditional for checkboxes

* Extracted conditional HTML into its own utility macro

* Removed `BaseController`

We don't know yet whether there will be any logic
to be shared in other potentiall controllers so
removing this for now.
`QuestionsController` inherits from `FormWizard.Controller`
directly now.

* questionPage.njk: Removed fictional `field.skip`

* Reworked renderField to be a normal partial

Instead of using macros. This should be slightly
nicer to use as it's a normal partial and solves
the problem of some of the used context not being
available.

* Moved `if field.component` logic into `renderField()`

No need for this to clutter the external template,
this is still part of the field rendering logic.

* Close <strong> element for unknown component warning