Fix tests to include answers ids
NOTE: In the test answers have locally unique IDs
(e.g. an answer ID may **not** be unique across
the various config or even within the same config).

This is not that important in the context of this
test and in fact we may not care about the answers IDs
being unique as long as these IDs are useful to
build some unique string which allows to discriminate
between the submitted values. For example:

Question ('12345'): Which part were injured?
Answers:
  - Answer ('A1'): Neck
    - Comment required (input name "12345-A1-comment")
  - Answer ('A2'): Leg
    - Comment required (input name "12345-A2-comment")
Question ('6789'): Was it a serious injury?
Answers:
  - Answer ('A1'): Yes
    - Comment required (input name "6789-A1-comment")
  - Answer ('A2'): No
    - Comment required (input name "6789-A2-comment")