Removed `classes`/`labelClasses` from Field type
These are presentation concerns and should be handled
by the template and not be part of the Field definition
itself.

NOTE: This is likely to break the spike/WIP views
but it doesn't matter as we're either tidying them
up or removing/replacing these. It was working because
the field was passed whole as-is to the GOVUK Design
System's macros (via `callAsMacro()`)