Removed `value` from `Field` type
A Field represents a question and it doesn't have a
fixed value. The value is provided by the user when
they submit the form.

For components with items (e.g. radios/checkboxes)
the items would represent the answer and these would
**have** values because these are the fixed codes
associated to the specific answer.

For example:

```
'Do you like dogs?' (name: 'Q1')
  (X) 'Yes' (value: 'yes')
  (  ) 'No' (value: 'no')
```

when the user submits this the value submitted for
the input named 'Q1' would be 'yes'.

In other words
- Field ~~ question (has name but no value)
- FieldItem ~~ answer (no name but has value)