Removed `fieldset`/`hint` from Field type
The former is a presenation concern that should be
handled by the template. This was including CSS classes
and the inner value was always the same as label anyway.

The latter is not currently a need, we may re-introduce
later but there is also no way to automatically
generate hints for questions from the questionnaire
Config...Maybe we'll reintroduce something similar
if we need it for other forms but ultimately it's
a future problem.