Generate conditional fields + Improvement to `Field`/`FieldItem` types (#228)
* renderField panic when component is not supported

* Attempt at simplifying field properties: label

Removed "text" and "label" with nested text/classes.
This is the format used by GOVUK Design System components
but it includes styles (which should be in the template/CSS)

NOTE: Update to existing fields done separately.
      This will likely break the WIP templates as they're
      using `callAsMacro()` and pass the whole field as-is

* WIP fields: Update fields' label to be a simple string

Mainly to make typecheck happy.

NOTE: This will likely "break" the WIP templates but
it doesn't matter as they're going to go soon-ish
anyway.

* Removed unnecessary `fieldset` from generate fields

This was only "used" to display the question label
which is a presentation/template concern.

* Use answer IDs to generate conditional fields

* Removed `classes`/`labelClasses` from Field type

These are presentation concerns and should be handled
by the template and not be part of the Field definition
itself.

NOTE: This is likely to break the spike/WIP views
but it doesn't matter as we're either tidying them
up or removing/replacing these. It was working because
the field was passed whole as-is to the GOVUK Design
System's macros (via `callAsMacro()`)

* Removed `leadingZeros` from Field type

This is a presentation concern and shouldn't be part
of the Field definition itself, the template should
determine how date inputs are presented instead.

NOTE: This is likely to break the spike's/WIP pages
but we're replacing/removing these anyway.

* Field type: Removed commented property

* Removed `text` from `Field`/`FieldItem` types

This is serving the same purpose of `label` and it's
just noise and confusing to have 2 very similar properties.

NOTE: This is breaking existing spike/WIP templates.
But they're on their way out.

* Removed `value` from `Field` type

A Field represents a question and it doesn't have a
fixed value. The value is provided by the user when
they submit the form.

For components with items (e.g. radios/checkboxes)
the items would represent the answer and these would
**have** values because these are the fixed codes
associated to the specific answer.

For example:

```
'Do you like dogs?' (name: 'Q1')
  (X) 'Yes' (value: 'yes')
  (  ) 'No' (value: 'no')
```

when the user submits this the value submitted for
the input named 'Q1' would be 'yes'.

In other words
- Field ~~ question (has name but no value)
- FieldItem ~~ answer (no name but has value)

* Removed `fieldset`/`hint` from Field type

The former is a presenation concern that should be
handled by the template. This was including CSS classes
and the inner value was always the same as label anyway.

The latter is not currently a need, we may re-introduce
later but there is also no way to automatically
generate hints for questions from the questionnaire
Config...Maybe we'll reintroduce something similar
if we need it for other forms but ultimately it's
a future problem.

* Removed `hint` property from `FieldItem`

Similarly to questions, report type report type configuration
doesn't include information about hints for anwers.

This may be nice to have possibly maybe in the future
but ultimately it is a future problem.

* Removed `id` from Field type

Fields had both `name` (required) and `id`, the latter
was always matching the name so no point in keeping both
and creating confusion.

Notice that `FieldItem` (the answers) don't have a `name`
as they're value are submitted for the parent field `name`
input.

* Fixed tests still using `text` instead of `label`

* Generate conditional fields

Fields for `requiredDate`/`requiredComment` need
to be part of the fields in order for Form Wizard
to be aware of them/validate them/etc...

These are also included in the corresponding step
and ignored when rendering the top-level fields.

* Removed dashes in `{%-` and `-%}` for consistency

These means "remove whitespaces" but in out case
it doesn't matter and it's misleading/noise.

* Removed `/assaults` test page

This is broadly doing what `/:reportType/questions`
is doing but for the ASSAULT report type.