Introduced `NextCondition` type
This is the same list of valid values previously
inlined in the `NextStep` type and in theory it shouldn't
make any difference.

However I found that TS seemed to struggle to accept
a list of these coming out of `Array.prototype.map()`
unless I specified its type. I could have specified
`const next: FormWizard.NextStep` but this is an
array of conditions really so use the singular word
"NextStep" could be misleading.