Generated steps work with multiple values
When a question is allowed to have multiple values
submitting the form wasn't going to the next page.

This was because a field with `multiple: true` would
be submitted as `string[]`. This is correct but the
naive next step condition would still try to match
this using the default `===` op which in this case
doesn't make sense:

e.g. user submitted `['dog', 'fox']` but the value in
the condition is `'dog'`, so `'dog' !== ['dog', 'fox']`.

In order for the next step work the condition has to
check whether any of the submitted values matches
the condition's value:

For example:

```
// Pseudo-code
[
  { value: 'cat', next: 'q2', op: '???' },
  { value: 'dog', next: 'q2', op: '???' },
  { value: 'fox', next: 'q2', op: '???' },
]

valuesSubmitted = ['dog', 'cat']
```

in this case we want Form Wizard to check if `'cat'`
is included in `['dog', 'cat']`, if it is that's a
match and user would be directed to `q2`.

This is done using function to check submitted values
when field has `multiple: true` (otherwise usual
`===` operation applies).