IR-519: Various improvements to questions steps/fields (#234)
* Fix date conditional field JS error

`leadingZeros` is a boolean, not a string!

* On error, user choices are retained for checkboxes

* Steps don't include fields for inactive answers

This shouldn't be harmful but it's also unnecessary.
In practice it shouldn't make much difference in
actice report type config with very few inactive answer -
nevertheless it's best to ignore inactive answers.

* Add conditional fields only for active answers

* Don't generate fields for inactive questions

This shouldn't be harmful but it's wasterful nevertheless.

* Validate dates in conditional fields

* Introduced `NextCondition` type

This is the same list of valid values previously
inlined in the `NextStep` type and in theory it shouldn't
make any difference.

However I found that TS seemed to struggle to accept
a list of these coming out of `Array.prototype.map()`
unless I specified its type. I could have specified
`const next: FormWizard.NextStep` but this is an
array of conditions really so use the singular word
"NextStep" could be misleading.

* Removed redundant string from `NextStepCondition` type

This is `string |` is already mentioned in `NextStep`
and judjing from the wording in Form Wizard docs is
not technically a "condition" (it just goes to the given
path unconditionally):

> The next step for each step can be a relative path,
> an external URL, or an array of conditional next steps.

* Generated steps work with multiple values

When a question is allowed to have multiple values
submitting the form wasn't going to the next page.

This was because a field with `multiple: true` would
be submitted as `string[]`. This is correct but the
naive next step condition would still try to match
this using the default `===` op which in this case
doesn't make sense:

e.g. user submitted `['dog', 'fox']` but the value in
the condition is `'dog'`, so `'dog' !== ['dog', 'fox']`.

In order for the next step work the condition has to
check whether any of the submitted values matches
the condition's value:

For example:

```
// Pseudo-code
[
  { value: 'cat', next: 'q2', op: '???' },
  { value: 'dog', next: 'q2', op: '???' },
  { value: 'fox', next: 'q2', op: '???' },
]

valuesSubmitted = ['dog', 'cat']
```

in this case we want Form Wizard to check if `'cat'`
is included in `['dog', 'cat']`, if it is that's a
match and user would be directed to `q2`.

This is done using function to check submitted values
when field has `multiple: true` (otherwise usual
`===` operation applies).

* Added tests for `generateSteps()`/`generateFields()`

* Added tests for `checkMultipleValues()`

* Removed now obsolete TODO comment