Generated steps now group answers
Answers leading to the same next question would be grouped
together in the same next step condition.

This is more legible and possibly more compact.
Potentially for questions with a lot of answers
it means the `checkMultipleValues()` function (or `===`)
would be invoked less times.

It also doesn't harm that the format is more legible
and easier to debug.