Implementation of questions grouping
Contiguous can be potentially be merged into a single step.
The result is a page with multiple fields/questions in it.

This can happen if there is no branching (all answers lead to same next
question). The next step also has to be a step that is only reached by
the given question, e.g. the current next is the sole parent of the
next step. This means a question belongs only to a single group.

Furthermore, grouping is prevented if too many options would appear on
screen. A limit of 20 total answers per page is set.

This is particularly important because some of the questions can easily
have 10/20 possible answers and we wouldn't want to group 3-4 questions
and end up with 40 options on a single page as this would be overwhelming
for the user.

On the flip side, a series of contiguous questions with yes/no answers
would easily fit on a single page without much trouble.