IR-519: Automatically group contiguous questions into a page (#252)
* `next` in NextStepCondition can be `null`

This is the case when the given answer leads to
no further questions.

* Updated `generateSteps()` tests for steps grouping

Contiguous, non-branching steps are expected to be merged into
a single step.

* Implementation of questions grouping

Contiguous can be potentially be merged into a single step.
The result is a page with multiple fields/questions in it.

This can happen if there is no branching (all answers lead to same next
question). The next step also has to be a step that is only reached by
the given question, e.g. the current next is the sole parent of the
next step. This means a question belongs only to a single group.

Furthermore, grouping is prevented if too many options would appear on
screen. A limit of 20 total answers per page is set.

This is particularly important because some of the questions can easily
have 10/20 possible answers and we wouldn't want to group 3-4 questions
and end up with 40 options on a single page as this would be overwhelming
for the user.

On the flip side, a series of contiguous questions with yes/no answers
would easily fit on a single page without much trouble.

* Improvements to grouping logic

Added a test which generate the steps/does the
grouping on a non-trivial config (a copy of the
FINDS config).

Grouping now correctly prevent a page from having
more than 20 answers across its fields.

Moved/tidied some of the logic into smaller function
and simplified nested functions signature by using
parent function variables.

* More doc comments for the grouping utilities

* Use `has()` to check if a Set contains an item

* Tweaks to `countStepAnswers()`

`step.next` is already an array (of type `NextStepCondition[]`)
so `Object.values()` is redundant - although not harmful.

Removed `Object.values()` and moved type conversion
inside the loop block which is possibly more readable.