IR-68: Import default express namespace types using a name that is not itself exported as well (#257)
* Import default express namespace types using a name that is not itself exported as well
This is good practice according to eslint because
```typescript
import type Express from 'express'
// and
import type { Express } from 'express'
```
refer to different things. It’s best that we use the package name `express` as the default import name instead. This “problem” was spotted when testing migrating to eslint 9.

* Add missing express type imports
I think the previous version worked because of the ambient express type definitions in `/server/@types`