Questions controller/steps accept multiple values
This is because some questions may allow for multiple choices
(e.g. checkboxes) and this is to make TS typechecker happy.

In practice this means that the values could be normal
`string` (for radiobuttons for example) or `string[]`
in the case of questions that allow potential multiple
answers (e.g. checkboxes).