IR-576: View report answers (#269)
* Set back link in questions page

* QuestionsController reads report

* Generic questions route moved under `/reports/:id/`

URL contains the ID of the report to show questions for.
The config is read based on the report's type now.

Rest is the same.

* Questions router: Set `mergeParams = true`

This is necessary in order for middleware to be able
to read things like params (how the report ID is provided)

* Questions controller/steps accept multiple values

This is because some questions may allow for multiple choices
(e.g. checkboxes) and this is to make TS typechecker happy.

In practice this means that the values could be normal
`string` (for radiobuttons for example) or `string[]`
in the case of questions that allow potential multiple
answers (e.g. checkboxes).

* Populate form answers from API Report

Populate the questions form fields with values coming from the API report.
Fields are populated only when the user didn't already set these values,
e.g. user-submitted choices are retained.

Some mapping logic is needed to determine the various field names
from the Report API as these don't contain the answer IDs (the data coming
from the API has the list of `responses` and the `response` field is
the answer "code", e.g. the text which is used as an identifier).

Any given question may have multiple responses.
Each Response may also require comment and/or date.

For example a question 'question-1' may allow for multiple answers. Then
each answer may also optionally require a date and/or a comment, for example
`answer-A` may require both, these fields values (`question-1-answer-A-date`
and `question-1-answer-A-comment`) would be read from the `response`'s
`responseDate` and `additionalInformation` fields respectively.

But the important point is that several responses may need these, e.g.

Question 1
- [X] Answer A
  - Date:    [ 30/12/2024 ]
  - Comment: [ Answer A's comment ]
- [X] Answer B
  - Date:    [ 31/12/2024 ]
  - Comment: [ Answer B's comment ]

* Added tests for answers displaying logic

* Fix for dynamic import of Report type config

Dynamic import works differently depending on whether
they happen when running the transpiled application
or the TS directly (e.g. in tests)

This change account for these differences by detecting
the current file extension and importing the config
accordingly.

* Set questions' back link to `/reports` page

* Added note about removal of QID-stripping logic

API is being changed to remove these prefixes so we
should remove this logic from the UI once that's gone.

* Include report id in "question/answer not found" errors

* Log 'question/answer not found' as errors

They currently don't cause the UI to 500. They shouldn't
happen but they are significan enough that it makes
sense to raise them to 'error' level as they warrant
investigation if they ever happen.

* QuestionsController: Don't return value from callback

* `populateReport` also loads report type config

Made this logic a bit more generic so that the questions
router/controller don't need to load the report/config
twice.

* Question router/controller uses `populateReport`

Removed duplicaed logic to load report and its config
and use `populateReport` middleware instead.

* Questions route test: Added note

Added comment to explain report is updated in each
test.