PoC Ability to answer questions
This is making requests to the API to save the answers.
It's currently using the `POST /incident-reports/{reportId}/questions`
[endpoint] which is designed to **append** answers
rather than update them.

As a result, this is working in a way that makes
sense only when a given answer(s) is answered the
first time. Update currently doens't work as expected.
In that case the same question is showing multiple time
in the report's questions which is most definitly
what we'd want.

Furthermore the displaying of existing questions' answers
doesn't play well with this bad data as "the first answer
wins" and the rest is ignored.

Ideally, we'd want some sort of `PUT /incident-reports/{reportId}/questions`
endpoint which either add a question's response or replaces its responses
if it was already present in the report.

[endpoint]: https://incident-reporting-api-dev.hmpps.service.justice.gov.uk/swagger-ui/index.html#/Questions%20and%20responses%20forming%20an%20incident%20report/addQuestionWithResponses