Added utility to add "?" to questions
A lot of the questions in NOMIS were clearly questions but they
lacked a "?". This utility simply check for some common words used
to ask questions (e.g. "who", "what", etc...) and if the label
is detected to be a question adds the question mark if it's not already
present.

This utility is conservative and it doesn't adds a "?" unless the label
is a question.

labels like `''ESTIMATED COST OF DAMAGE'` or `'DESCRIBE THE TYPE OF SPRAY'`
would be left alone.

However by inspection of several of the report configurations this should
correct most of the missing question marks.