Questions controller reduce errors (#286)
* Refactor: Set steps/fields in `populateReport()`

This is so that these are available from the
`QuestionsController` as well.

* Don't look up conditional fields as questions

This is to avoid unnecessary false errors in the logs.

When questions/answers are submitted the submitted
values' keys are fields names. However these field
names can be either a question ID or a conditional
field name (e.g. comment/date)

This change only look up question IDs in the report
config.

NOTE: This does not change what we post to the API.
Conditional fields values are still submitted when
the question's answers require date/comment.