Sentence case: Split by non-word boundaries
This solves a lot of issues were an acronym's case was
changed to lower case when not surrounded by space.
This lead to inconsistencies in the case of acronyms.
Sometimes even in the same sentence, e.g.
`Has the NOMS SFO team been informed (sfo cases only)?`

This is an improvement as a lot of the words we want
to preserve the case for are often surrounded by non-spaces
like `/` or parenthesis.

Unfortunately it's a tradeoff and this change means
some preserved words (e.g. `N/A`) which contains
some of the non-space characters are now not preserved.

However these words are not as frequent and usually
this doesn't look too bad and most importantly shouldn't
make them ambiguous. So overall this improves the case
of many of the questions/answers.

We could try and iterate further to try and improve
the case for these edge cases but it's possibly
excessive.